+++ apps/src/sketchlab/reactFlow/nodes/nodeToolbars/ShapeNodeToolbar.tsx
@@ -3,6 +3,7 @@ import React from 'react';
 import FontSizeGroup from './FontSizeGroup';
 import HandleVisibilityToggle from './HandleVisibilityToggle';
 import NodeToolbarShell from './NodeToolbarShell';
+import SwatchGroup from './SwatchGroup';
 import {
   BACKGROUND_PALETTE,
   DEFAULT_BACKGROUND_COLOR,
@@ -10,8 +11,7 @@ import {
   DEFAULT_FONT_SIZE,
   DEFAULT_STROKE_COLOR,
   STROKE_FONT_PALETTE,
-} from './shapePalettes';
-import SwatchGroup from './SwatchGroup';
+} from './toolbarPalettes';
 import {useNodeToolbarData} from './useNodeToolbarData';
 
 interface ShapeNodeToolbarProps {
